fix(indexed_enums): use int16 for enum indices to support >256 members
Enum index arrays were encoded as uint8 (0-255). Enums with more than 256
members (e.g. NAF with 733 codes) had indices truncated (e.g. 555 -> 43),
causing wrong values in simulations.

- types.py: use ENUM_ARRAY_DTYPE (int16) instead of uint8 for EnumDType
- _utils.py: doctest expected output updated to int16
- _enum_type.py: doctest expected output updated to int16

Variable config already uses ENUM_ARRAY_DTYPE for storage; encoding now
matches, so no overflow occurs.

+++ openfisca_core/indexed_enums/_enum_type.py
@@ -26,7 +26,7 @@ class EnumType(t.EnumType):
         ...     TENANT = "Tenant"
 
         >>> Housing.indices
-        array([0, 1], dtype=uint8)
+        array([0, 1], dtype=int16)
 
         >>> Housing.names
         array(['OWNER', 'TENANT'], dtype='<U6')
